refactor: move logic for creating into the service (#18)
* Remove unused imports

* Create InsertFields subclass

* Create new method

* Use Insertfields
